+++ services/DailyServices.qml
@@ -5,6 +5,7 @@ import qs.services.brightness
 import qs.services.connectivity
 import qs.services.media
 import qs.services.power
+import qs.services.weather
 
 Scope {
     readonly property var audioService: AudioService
@@ -12,4 +13,5 @@ Scope {
     readonly property var connectivityService: ConnectivityService
     readonly property var mediaService: MediaService
     readonly property var powerService: PowerService
+    readonly property var weatherService: WeatherService
 }